BJS2-40891 - Jacoco для post_service
