+++ src/main/java/faang/school/postservice/dto/like/LikeDto.java
@@ -1,5 +1,7 @@
 package faang.school.postservice.dto.like;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Data;
@@ -16,5 +18,7 @@ public class LikeDto {
 
     private Long commentId;
     private Long postId;
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm")
     private LocalDateTime createdAt;
 }
