--- src/main/java/faang/school/postservice/filter/album/AlbumFilter.java
@@ -1,13 +1,0 @@
-package faang.school.postservice.filter.album;
-
-import faang.school.postservice.dto.album.AlbumFilterDto;
-import faang.school.postservice.model.Album;
-
-import java.util.stream.Stream;
-
-public interface AlbumFilter {
-
-    boolean isApplicable(AlbumFilterDto filters);
-
-    Stream<Album> apply(Stream<Album> albums, AlbumFilterDto filters);
-}
