+++ src/main/java/faang/school/postservice/mapper/like/LikeMapper.java
@@ -1,0 +1,31 @@
+package faang.school.postservice.mapper.like;
+
+import faang.school.postservice.dto.like.LikeDto;
+import faang.school.postservice.model.Like;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.NullValueCheckStrategy;
+import org.mapstruct.NullValuePropertyMappingStrategy;
+import org.mapstruct.ReportingPolicy;
+
+@Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
+public interface LikeMapper {
+
+    @Mapping(source = "id", target = "id")
+    @Mapping(source = "commentId", target = "comment.id",
+            nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE,
+            nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS)
+    @Mapping(source = "postId", target = "post.id",
+            nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE,
+            nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS)
+    Like toEntity(LikeDto likeDto);
+
+    @Mapping(source = "id", target = "id")
+    @Mapping(source = "comment.id", target = "commentId",
+            nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE,
+            nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS)
+    @Mapping(source = "post.id", target = "postId",
+            nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE,
+            nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS)
+    LikeDto toDto(Like like);
+}
