+++ src/main/java/faang/school/postservice/service/album/AlbumService.java
@@ -3,13 +3,13 @@ package faang.school.postservice.service.album;
 import faang.school.postservice.dto.album.AlbumDto;
 import faang.school.postservice.dto.album.AlbumFilterDto;
 import faang.school.postservice.exception.AlbumException;
-import faang.school.postservice.filter.album.AlbumFilter;
-import faang.school.postservice.mapper.AlbumMapper;
+import faang.school.postservice.filters.album.AlbumFilter;
+import faang.school.postservice.mapper.album.AlbumMapper;
 import faang.school.postservice.model.Album;
 import faang.school.postservice.model.Post;
 import faang.school.postservice.repository.AlbumRepository;
-import faang.school.postservice.service.PostService;
-import faang.school.postservice.validator.AlbumValidator;
+import faang.school.postservice.service.post.PostService;
+import faang.school.postservice.validator.album.AlbumValidator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
