+++ src/main/resources/application.yaml
@@ -47,3 +47,6 @@ moderation:
   cron: 0 0 0 * * *
   zone: Europe/Moscow
   thread-pool-size: 1
+
+api:
+  version: "/api/v1"
